requirements.txt
PyYAML==6.0.2
streamlit==1.42.1